rag/requirements.txt
asgiref==3.8.1
certifi==2025.6.15
charset-normalizer==3.4.2
Django==5.2.3
djangorestframework==3.16.0
dotenv==0.9.9
drf-yasg==1.21.10
idna==3.10
inflection==0.5.1
joblib==1.5.1
numpy==2.3.1
packaging==25.0
psycopg2-binary==2.9.10
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.4
scikit-learn==1.7.0
scipy==1.16.0
sqlparse==0.5.3
threadpoolctl==3.6.0
uritemplate==4.2.0
urllib3==2.5.0
